xlang: Single goroutine which owns shutdown and cleanup of connection (#10615)
* xlang: Single goroutine which owns shutdown and cleanup of connection

Previously we had several places do a mixture of shutdown messages and closing
the connection. This lead to fun race conditions which had nil panics. For
example we add a connction to our list before `c.conn` is non-nil. So if that
fails instantly we then call shutdown which leads to accessing a nil c.conn to
send the shutdown messages.

Instead we do what we did previously for resource cleanup, and move all of
this to a goroutine which responsible for cleanup and only exists if we
successfully init.

* log jsonrpc2 errors when shutting down