pkg/vcs/gitcmd,graphqlbackend: fix diff search behavior when timeout hit
Previously, in RawLogDiffSearch, if the 2nd command (`git show`) took a long time, it could exhaust the entire remaining timeout before the deadline. Then the `git rev-parse` calls in `filterAndResolveRefs` would immediately time out, and an error would be returned. This commit fixes that bug by reserving 20% of the remaining time for the `git rev-parse` commands. This is a heuristic similar to the others in use for dividing the total time between `git log` and `git show` (see comments in the RawLogDiffSearch impl).

This commit also fixes another bug where a `context.DeadlineExceeded` error returned by RawLogDiffSearch would be ignored in `searchCommitDiffsInRepos`, thereby not reporting that the call timed out.

fix #10647
fix #10281

(although note there are some edge cases where those issues will still occur, but they are now much rarer after this commit, and they will be reported correctly as timed out instead of as "no results" now)