graphqlbackend: don't ignore results from context-canceled searches
Previously we could assume that `if ctx.Err() != nil {` then our search request was canceled. This assumption no longer is correct when the context has a timeout (as is the case with the new `timeout:` operator and the search timeout feature flag). It assumes the only reason why `ctx.Err() != nil` could be true is if the `handleRepoSearchResult` func call previously returned a `fatalErr != nil` and therefore we called `cancel()`. But now `ctx.Err() != nil` can also be true for other reasons, such as if the context timed out due to the overall search timeout.

This causes timed out repos (in text and diff search) to be silently ignored and not reported as
having timed out. The user would see "no results" not "N repos timed out".

There are other bugs in diff search that must be fixed before timeouts are truly communicated
correctly (see https://github.com/sourcegraph/sourcegraph/pull/10665), but this commit is a prereq
to fixing those.

fix #10634