cmd/frontend/internal/pkg/langservers: use "docker restart" instead of stop & start
Previously we performed a restart naively by using `docker stop <container>`
and `docker start <container>`. This was problematic because there is a small
time period in which the `docker start <container>` command is running but the
old container has not yet stopped. i.e. `docker stop <container>` does not
guarantee the container is stopped by the time it exits.

However, `docker restart <container>` does guarantee it will restart the
container, which is exactly what we want.

Fixes #10716