xlang: Avoid panic when Close is called (#10742)
We call close on done more than once if the connection goes away, leading to a
panic in lsp-proxy (you can't close a closed channel). The first close is done
in our DisconnectNotify listener, the second when the connection is removed
from the server proxy list (via `serverProxyConn.Close`). We don't actually
need to close in the DisconnectNotify, so we just remove that bit of logic. We
add extra protection with a `sync.Once` around closing done.