xlang: Handle missing repos (#10729)
repoURL is returned from the language server based on what it finds in the
repository. S we may have not previously seen the repository or the repo
itself may not exist. As such we add code to skip over repos we haven't seen.

For sourcegraph.com we will only skip repos which are not cloned, for other
instances (depending on settings) we will instead use a later federated xdef
call.