cmd/frontend/internal/graphqlbackend: properly indicate that you cannot restart langserver unless it is enabled
In the past, it was possible to restart a langserver that was not enabled
because it was simply the action of stop (noop) and start. i.e., restarting a
langserver that was not enabled would simply start it.

This behavior was confusing, and was made no longer possible after we switched
to `docker restart <container>` for restart logic rather than `stop` and
`start`, but this code was never updated to reflect that so the frontend still
showed the restart button prior to this change.

Fixes #10751