web: site-admin: always show langserver update button when possible (even if langserver disabled)
Prior to this change we did not show the langserver update button when the
langserver was disabled, but we did show it when the langserver was in the
'neither enabled nor disabled' state. This was inconsistent.

I chose to always display this button, rather than only showing it when the
langserver is enabled, because we always have this action present AND it is
reasonable to think that a user may want to update a langserver before starting
it in the event of e.g. there being a langserver with a system performance
degradation issue that we've fixed.

Finishes #10751