cmd/frontend/internal: set Vary headers more granularly based on actual auth middleware in use
Previously we only set Vary in one place, in the AuthorizationMiddleware. This happened to be called in all cases where it was necessary, so the correct Vary header was always sent in the HTTP response. If, however, a code change (as will be made in a subsequent commit) removed the AuthorizationMiddleware from a request path that needs to vary on Cookie or Authorization, then the Vary header would not exist. This commit makes it so the Vary header is added on an as-needed basis by all middlewares that perform auth and need the Vary header to be set, to avoid a security issue if a mistake is made in the future.

This causes some HTTP responses to have multiple Vary headers with some duplicate values. This is acceptable; all Vary headers' values will be merged. See RFC 7231 https://tools.ietf.org/html/rfc7231#section-7.1.4 and for merging behavior, RFC 7230 https://tools.ietf.org/html/rfc7230#section-3.2.2.