cmd/frontend: require cookie auth for SAML and OpenID handlers
Previously, they would both accept a session cookie value that was actually provided in the HTTP request Authorization header (not the Cookie header). This was because the web client frontend sent auth in the Authorization header.

This commit makes the web client frontend use cookies for auth (and send X-Requested-By, which the HTTP API requires to be present in order to perform cookie auth to avoid CSRF). This means we can remove the ability to use the HTTP request Authorization header for auth.

Note that this is a slight temporary regression; any API clients that authed to the Sourcegraph API using the Authorization header would break when the auth provider is SAML or OpenID. This affects a few internal Sourcegraph tools but no known customer or user tools in use by 3rd parties. A subsequent commit will add back the ability for API clients to authenticate by providing an access token in the Authorization request header.