cmd/frontend/internal/cli: use HTTP API/app auth middleware when mounting/wrapping handlers for clarity
This moves where the HTTP API/app auth middlewares are applied, from the NewXyzHandler funcs to the place where the handlers are actually mounted. This improves separation of concerns in the NewXyzHandler funcs (letting them deal solely with actors for auth, not HTTP headers as well) at the expense of less encapsulation (it is now theoretically possible to accidentally create and mount an API HTTP handler elsewhere that does not check for auth). This tradeoff is worth it.