cmd/frontend,web: move GraphiQL to app router to expose it
It should be visible to any cookie-authed users, which is more similar to the behavior of app pages (which accept cookie auth) than API handlers (which generally don't accept cookie auth, except in certain cases).

Moving it to an app handler lets us get this desired behavior without poking holes in the API router just for this page (which would increase code complexity and increase the risk for a vuln).