search: support @username expansion in diff/commit search author: and committer: filters (#10906)
This makes it so you can use `@username` references in `author:` and `committer:` search filters in diff/commit search. They will be expanded to match all known and verified emails for the user.

That is, `author:@sqs` is equivalent to `author:sqs@sourcegraph.com author:qslack@qslack.com author:qslack@gmail.com` assuming those are the emails known to Sourcegraph (and verified) for the user `@sqs`.

Limitations:

- only supports mentions of existing Sourcegraph user accounts (does not query GitHub/SSO/etc. for users)
- only supports expanding to email addresses that were added to Sourcegraph (ie in the user settings emails page) and verified

Because of these limitations, we will not announce or expose this feature yet. It is very unlikely a user would come across this feature unexpectedly and have it exhibit undesired behavior. (If a user searches for, eg, `author:@foo` and there is no Sourcegraph user account named `foo`, it will search for the literal regexp `@foo`.)

Motivation: gets part of the way to a fix for #10699