pkg/env: correct inverted log level comparison
Prior to this change, the log level comparison was inverted which meant that
anything logged to e.g.:

```
	// DebugOut is os.Stderr if LogLevel includes dbug
	DebugOut io.Writer
```

Would be logged even if `SRC_LOG_LEVEL=crit`. See https://github.com/sourcegraph/sourcegraph/issues/10940#issuecomment-385098888

After this change, the correct / intended behavior is in place: `DebugOut` is logged to only if `SRC_LOG_LEVEL` (`lvl`) is inclusive of `debug` (`maxLvl`). Not logged in the case of `crit=0` `error=1` `warn=2` or `info=3`, which all have a value less than `debug=5`.

Fixes #10940