db: always set site initialized when creating a user account
Previously, it was possible to create the initial site admin account and NOT initialize the site. This left the site in a bad state, where it would show the site initialization screen but submitting that form would fail (#9885 #10932).

The cause of the problem was a bug in `(*Users).Create`: it tried to `UPDATE site_config` assuming that it contained at least one row, but if the user signed in via SSO (SAML/OpenID/etc.), it's possible for the table to have zero rows at this point.

helps with #9885 #10932