add migration/UI to fix bad site initialization state when user accounts exist
This adds a DB migration and UI notice to handle the case where the site is not initialized but users exist. It was possible to get into this state if the first user was from SSO (OpenID/SAML/etc.). After this commit, it's still possible to get into this state if the admin manually deletes/updates the `site_config` table (which users have been known to do).

helps with #10932