cmd/frontend/internal/auth: better error message when user logs in after auth provider has changed
If Alice logs in via OpenID/SAML/builtin as `alice` and then the auth provider changes and she logs in via the new auth provider, she will get an error because we only allow logins via the account's original auth provider. This is a limitation we would like to remove eventually. For now, this improves the error message displayed in this case.