cmd/frontend/internal/db: delete & invalidate an access token when its creator or subject is deleted (#10984)
Previously, if the creator or subject user of an access token was deleted, the access token would remain valid. (This was because the db.AccessTokens store did not check the `users.deleted_at` column, which indicates whether the user is deleted, because we use [soft deletes](https://sourcegraph.sgdev.org/github.com/sourcegraph/sourcegraph/-/blob/docs/storage.md#hard-vs-soft-deletes).)

Now, it checks that both users are present and not deleted, and it also sets the `deleted_at` column for all tokens associated with a user when the user is deleted.

It is possible for some tokens (whose creator/subject user was deleted prior to this commit) to still be in an inconsistent state. After this commit, those tokens would no longer be valid, but they would be displayed to site admins in the access tokens page. The number of such tokens is likely so small (and the negative consequences are so small, just some extra noise on the site admin area) that it is not worth fixing.