graphqlbackend: Check for empty repository in DefaultBranch (#10998)
DefaultBranch is documented to handle empty repositories. However, the `HEAD`
reference always exists so it would not fail. So in an empty repository it would
just point to a ref that doesn't actually exist. As such we ensure that `HEAD`
points to an actual object.

This likely fixes a few bugs for code which relies on DefaultBranch. One such
bug is displaying the status of empty repositories in indexed search.