graphqlbackend: Translate zoekt HEAD to default branch (#10997)
Zoekt now just uses `HEAD` as the branch name it indexes. The graphql indexed
search status had an assumption that it would match the actual default branch
name. As such we now translate `HEAD` into the default branch name.

Another fix for this is to instead get zoekt to use the default branch name
rather than just HEAD. However, this is the simpler approach.