PersonLink: make it apparent that the account-link is a secondary feature
Eliminates confusion in #10699 where the contributor was displayed as a user account, but the row's "N commits" link only filtered by one of the author's email addresses.

The true solution is to group all contributor rows that correspond to the same user account, but that takes more work (and needs some product thinking, since it duplicates the role of mailmap).

Fixes #10699