access tokens: add concept of scopes (only 1 is supported now, "user:all"; all new/existing tokens have it) (#10996)
This is a precursor to adding a "site-admin:sudo" scope to access tokens that lets the holder assume the identity of any other user while performing an action.

All existing access tokens are migrated to have the single scope "user:all". No other scopes are permitted, and the GraphQL API does not let you create an access token with no scopes (to avoid users thinking that a token with no scopes is "safe" to share).