implement access token site-admin:sudo scope (allows token holder to perform action as any other user) (#10994)
Adds a new access token scope "site-admin:sudo". Only site admins may create and be the subject of these tokens.

When an HTTP request contains a header as in the following example:

```
Authorization: token-sudo token="mytoken",user="alice"
```

the request context's actor is set to "alice", regardless of who created the token or issued the request. (If the token "mytoken" doesn't have the site-admin:sudo scope, or if the subject isn't an admin, or if "alice" is not the username of an existing user, the request will fail.)

Motivation: This supports the use case of an external service (eg Phabricator) that wants to provide Sourcegraph functionality dependent on a Sourcegraph site that requires authentication, without needing each user to individually log in. The external service can hold a Sourcegraph sudo access token and send requests to Sourcegraph's API using the sudo token, authenticated as the user who is accessing the external service.