use new JSON Schema codegen tool (github.com/sourcegraph/go-jsonschema) (#11060)
Right now this tool emits the same Go code for our JSON Schemas (minus tiny and insignificant changes you can see in this commit). So it is safe to switch to.

In the future, this new tool/library will make it easier to support default values, union types, and more. (I was not able to find an existing tool/library that I trusted to solve our needs here.)