pkg/conf: more accurate AuthProvider that reflects existence of legacy config
Previously, the following site config would be reported as having `conf.AuthProvider() == "foo"` and `conf.AuthSAML() != nil`:

```
{
  "auth.provider": "foo",
  "auth.saml": { ... }
}
```

This did not actually cause any problems, because we checked for `conf.AuthSAML()` (and the analogs for OpenID Connect and HTTP header auth) where it mattered. However, it led to the possibility of confusion, or of serious security mistakes if the `conf.AuthProvider()` value was trusted (e.g., if it caused "builtin" auth to be used when the site admin intended to use SAML).

The new code gives top priority to SAML or OpenID Connect auth specified in the site config in any way: either by `auth.provider` or by setting the legacy properties (such as `samlSPCert`, `oidcClientSecret`, etc.). This ensures that no site is DOWNGRADED from SAML/OpenID to builtin auth as a result of this commit.