auth: move override token/secret to its own auth provider, extricate it from the OIDC auth provider
It applies to all auth providers and is not OIDC-specific, so it never belonged here anyway.

In a future infra commit, I will switch our dogfood config to use the new OVERRIDE_AUTH_SECRET env var (it is backcompat with the old config so there will be no interruption in our e2e tests).