server: Shutdown if any process dies (#11134)
If any process managed by server dies, we shutdown all processes. If something
has died, then the container may be unresponsive or in a degraded state. As
such we need to recover. We could try to recover via starting the process
again. But to do this robustly requires many changes, and likely operators
would like control over it. Rather, we move lifecycle management up to the
container orchestrator. Docker and Kubernetes both have robust methods to
responding to a container dieing.