cmd/frontend/internal/auth: always wrap HTTP handlers with http-header middleware (#11123)
Supports auth middleware that is always wrapped and that chooses whether to enable/disable itself per-request by consulting config. This is a better design for auth middlewares because it (when all middlewares are migrated to adhere to it) allows for applying the new config without reloading and multiple auth providers.

In this commit, only the http-header middleware is migrated to adhere to the new design.

The only behavior change in this commit is that the auth.userIdentityHTTPHeader site config property does not require a restart to be applied when it is changed.