ineffectively conceal disabled repos from non-site-admin users (#11163)
Slightly helps with #10862.

This does NOT YET PROVIDE a guarantee that non-site-admin users can't see info about (or existence of) disabled repos. There are almost certainly ways to access such data, but they are no longer extremely obvious. This is security through obscurity and we will strengthen this guarantee in the future.