cmd/frontend/internal/auth: always wrap HTTP handlers with require-auth middleware (#11126)
Migrates the require-auth middleware to always wrap the handler (and selectively enable/disable itself per-request and based on the current site config). This design is more flexible than the previous one (which we're migrating away from) that selected the middlewares based on the site config as of server startup.

The only change in behavior in this commit is that the `auth.public` site config setting does not require a restart when changed.