auth: refactor OpenID Connect auth middleware to be enabled per-request (#11127)
This also means that if auth.provider is changed to "openidconnect", the change will take effect immediately. This is not mentioned in the changelog yet; I will wait for all auth providers to be updated to not require restart, and then I'll add a single bullet covering all of them to the changelog.