config: add auth.providers site config property, support reading auth provider from it (#11130)
* auth: forbid all access when auth provider is not set

This is a safer default than defaulting to the builtin auth provider.

The goal is that making a typo in your site config is highly unlikely to revert your site back to using the builtin auth provider (which may be less secure for your needs than SAML/OIDC/whatever).

* config: add auth.providers site config property, support reading auth provider from it

The new auth.providers site config is an array of auth providers (currently length 0 or 1). If an auth provider is specified in auth.providers, it takes precedence over the auth provider specified in auth.provider (and other auth site config properties).

Only 1 auth provider is supported; that limit will be lifted in the future.

This gets us toward being able to support multiple auth providers (and grouping their configuration together in a better way).