cmd/frontend/internal/cli: run ForbidAllRequestsMiddleware sooner
This is a fix to an edge case bug (not security-related issue) where if the user has an invalid cookie or access token AND the site has no auth provider configured, the client will see the cookie/access token error instead of the forbidden error (due to no auth provider being configured). It is slightly better to show the forbidden error because it would occur even if the access token and cookie errors were fixed.