cmd/frontend/internal/cli: run httpheader.Middleware sooner so it can ignore session cookies
When using the http-header auth provider, any session cookies set previously by previously configured auth providers should not be respected. To implement this behavior, httpheader.Middleware must run BEFORE session.CookieMiddleware and clear the cookie.

This most apparent behavior change here is that if you're logged in as a user and then the site's configured auth provider changes to http-header, you'll authenticate using only the http-header and not using your old session cookie. It is a rare case and the old behavior did not present any significant security issues, but this behavior is more correct and consistent with site admin intent.