session: DeleteSession always deletes HTTP session cookie even if the session store failed
If (e.g.) Redis is unreachable and we call DeleteSession (e.g., because the user logs out), we want the HTTP session cookie to be deleted. Previously the cookie would persist if the session store was unreachable.

Note that this does NOT make it so the cookie is deleted if the session store is unreachable during any request. It only applies to requests that call DeleteSession.