Handle multiple (possibly-empty) revspecs
In the existing code, each repo is matched against every pattern
until it matches one, then that pattern's revspec (if any) is used.
So if you have
	repo:foo repo:foo@a
the "@a" has no effect; the first one matches sooner, and its revspec
list is used.

To fix this, we have to distinguish between "user intentionally
specified a possibly-empty revision spec" and "user did not specify".
The intent is that, if you have multiple matching repo tokens,
and only one provides a revspec, that revspec should be used.

If multiple tokens specify revspecs, we want to intersect them, the
same way we'd intersect anything else in multiple tokens. If the
intersection is empty, that's an error

Expected outcomes:
    repo:foo repo:.*o@a
    	=> selects foo@a
    repo:foo@a repo:foo@a:b
    	=> selects foo@a
    repo:foo@a repo:foo|bar@a
    	=> also selects foo@a
    repo:foo@a repo:foo@b
	=> error indicating in some way that you specified conflicting
	   revspecs.

The reporting is a bit weird if someone writes a plain @, because
that counts as "having a refspec", but the refspec doesn't look
like anything.

Also modify test cases (two test cases internally relied on
parseRepositoryRevisions returning a non-empty list when no @ was
present, but so far as I know the actual functionality of the
product isn't affected), and add new test cases for handling multiple
repo specs which may or may not have @s.