migrations: retroactively fix downmigration for case when user has no emails (#11299)
After the 1514693059_user_emails_table migration, it became possible for users to have zero associated email addresses. This is OK. However, this downmigration did not properly handle that case. It would attempt to write a null value into a not-null column.

This workaround means that downmigrating past this version will add in dummy unroutable emails for users who had no associated email addresses. This is a very rare case, so this is acceptable.