add new experimental SAML auth provider/middleware (enhancedSAML experiment) (#11207)
* saml: make getActorFromSAML less implementation-specific

* saml: rename current SAML impl from xyz => xyz1 (to prepare for a 2nd impl)

* saml: extract readProviderConfig

* saml: clean up error messages

* saml: clean up tests

* saml: extract HTTP IdP metadata fetching and parsing

This code will be shared with the new SAML impl.

* saml: add enhancedSAML experiment (new SAML impl)

The experimentalFeatures.enhancedSAML feature flag enables the use of a new SAML auth provider and middleware implementation that is more architecturally similar to our current OpenID Connect impl. It uses the sg-session to store the authenticated actor instead of managing its own JWT (as the old SAML impl does).

The old SAML implementation (which was renamed to "xyz1", like "cache1", in the saml package) is still enabled by default and virtually unchanged by this commit.

* saml: extract and move redirectToAuthURL

* saml: refactor GET router

* fix race conditions in saml/openidconnect provider caches

Also adds 2 new tests. These tests failed with the race detector enabled (due to a race condition from the race detector, and also the t.Errorf) before the changes were made to the impl.