migrations: retroactively fix *another* downmigration for case when user has no emails (#11313)
Fixes a mistake in a downmigration.

After the 1514693059_user_emails_table migration, it became possible for users to have zero associated email addresses. This is OK. However, this downmigration did not properly handle that case. It would result in multiple users having the same `external_id` value of `native:null`.

Similar to 0db77cb14a5049efc69ae13b9417017f9767037b

This workaround means that downmigrating past this version will add in dummy unroutable emails for users who had no associated email addresses. This is a very rare case, so this is acceptable.