graphqlbackend: Avoid EnsureRevision when checking phabricator refs (#11319)
Using the default method to ResolveRevision results in gitserver ensuring the
ref exists. If the ref doesn't exist, gitserver will try and fetch it from the
remote host. However, the ref we want is created by us so won't be present on
the remote host. This coupled with clone limiters leads to the resolve rev
being very slow.

Instead we can use the same method used by search and other places. That is to
disable the fetch behaviour.