auth,config: handle slice of auth providers, not just a single auth provider (#11330)
This commit updates all locations where the auth providers config is read (`conf.AuthProvider()`) to consult a *slice* of auth providers, not just a single auth provider.

For example, the `conf.AuthAllowSignup()` helper used to work by checking the singular auth provider to see if it was `builtin` and had `allowSignup == true`. Now it iterates over the slice of all auth providers and returns true if any pass. In practice, these are equivalent, because the JSON Schema still limits `auth.providers` to having a single element, and the rest of the app still lacks support for multiple auth providers.

This commit is in preparation for full-featured support of multiple auth providers. In the meantime, the motivation of this commit is to update callers so they are already ready for the multiple-auth-provider world.