auth: implement backend and DB for a user's (multiple) connected accounts (#11329)
Previously a user could have at most 1 external account (SAML/OpenID Connect/etc.). A previous commit added the API and UI for displaying multiple connected accounts but still read the data in the old schema, which supported just a single account.

This commit adds a new DB table `user_external_accounts` and associated DB store to support having multiple external accounts. It also improves the handling of this case in the auth providers.

This functionality by itself is not very useful, but it is an important part of the foundation for deeper integrations with other tools and better configurability of user authn.