gitserver: Robust handling of origin url (#11357)
Some calls do not specify the remote URL, which means gitserver needs an
alternative way to know how to update the repository if it is triggered by such
a request. As such we store the last known URL as the origin remote (this used
to be the source of truth).

This commit cleans up our logic around this. Changes:
* Some calls may not have normalized the repo dir, leading to a does not exist
  error.
* Do not respect context timeouts. The git commands we run are _very_ quick and
  can mutate state. So we rather not kill them. These errors also often appeared
  as logs. Special casing context would be more verbose.
* Only call set-url if the url on disk is different.
* Handle the remote origin not existing.