remove old SAML implementation (#11422)
* remove old SAML implementation

The new SAML implementation uses a different SAML lib that I judge as high-quality and that allows us to implement SAML in a way very similar to OIDC. This is good.

Usually we would keep the feature flag around for backcompat. In this case, I judged the risk of regressing on old SAML functionality and security as high, if we had 2 SAML implementations. So in this case I made a judgment call to remove the old SAML code and not keep it in behind a feature flag.

* saml: remove "2" suffix from code that was previously in the new SAML impl

* mention SAML changes in changelog