SOURCEGRAPH_EXPAND_CONFIG_VARS: expand ${var} or $var in site config based on env vars (#11431)
Fix #10904 (a very valid request from Terin at Cloudflare). See Terin's email quoted there for the motivation for this change. More details: He would like to use a script that acquires the Bitbucket secret and supplies it to Sourcegraph in `docker run -e ...` because (1) the secret changes occasionally and he wants to be able to specify it in just a single scriptable place and (2) he doesn't want to use `sed` or similar to modify the JSON file before providing it to Sourcegraph because that would be annoying.

Experimental env var expansion in site config JSON: set `SOURCEGRAPH_EXPAND_CONFIG_VARS=1` to replace `${var}` or `$var` based on environment variables. Use `SRC_LOG_LEVEL=dbug` to see which variables were expanded.

If this works reliably and is useful, we can consider having it always enabled. Until further notice, only Cloudflare will try this.