auth: add ClientID to external account spec
This distinguishes between multiple auth providers that represent the same service but with different parameters, such as two OpenID Connect auth providers with the same issuer and different client IDs. It is better to represent the service ID and client ID in separate fields because that lets each field contain meaningful values (not just synthesized composite values) in more cases.