searcher: Handle exclusion ranges in lowerRegexpASCII (#11156)
An exclusion class is something like `[^A-Z]`. We need to specially handle it
since the user intention of `[^A-Z]` should map to `[^a-z]`. If we use the
normal mapping logic, we will do nothing since `[a-z]` is in `[^A-Z]`.

Unfortunately this resulted in some pretty hairy logic. This problem maps to
manipulating line intersections. There is possibly a way to write this in a
clearer way, but for now we have many more unit tests.