session: fix the failsafe code path for clearing session after error (#11442)
> This PR does not need to update the CHANGELOG because **it does not change any behavior that is observable by users**

If an unhandled error occurs due to the session cookie's value, we want to wipe it out so we can start over with a clean session. Unhandled errors can happen when you previously were running a different Sourcegraph instance on the same hostname, if you wipe your Redis cache, or if we have some other bug. In all cases, a forced logout and needing to sign in again is better than not being able to sign in at all.

fix #11432