gitserver: Do not cancel fetches if user request is canceled (#11447)
* gitserver: Do not block context in doRepoUpdate

The context for doRepoUpdate is not respected in doRepoUpdate2, since it may not
be the same context.

* gitserver: Use a background context for git fetch

Previously `git fetch` would be interrupted by the user cancelling the request
(which we aggressively propogate). We don't want this to happen since the fetch
is shared amongst requests. doRepoUpdate2 is only called from doRepoUpdate, and
that already correctly respects deadlines even if doRepoUpdate2 executes for
longer.