Un-bitrot some of the dev docs. (#11328)
* storage.md: update postgresql docs

In particular, on Linux, it's at least possible for postgres
to be configured such that it tries to use peer authentication
on a local unix-domain socket to connect.

Side note: I don't know whether this is performance-critical,
but local unix-domain sockets should be much faster than TCP.

No CHANGELOG entry because this is just internal documentation.

* Rework dev startup/configuration docs a bit.

docs/dev.md had some bitrot, such as references to a "src" binary
that appears not to be a thing anymore, and some of the instructions
were out of order, such as telling you to `npm login` in the sourcegraph
directory before telling you to clone sourcegraph.

Also cleaned up and organized the possible problems and troubleshooting
a bit.

Then got a bunch of comments in review and fixed things up further
and removed at least two new errors I'd introduced. And fixed a couple
more things.