gitserver: EnsureRevision checks existance on absolute commits (#11444)
This is a pretty bad bug and I'm surprised it hasn't bit us before. Its likely
that we don't notice on dogfood (low traffic) and sourcegraph (high concurrent
updates), so the normal update method works.

Previously if EnsureRevision was an absolute commit ID it wouldn't actually
check if the revision existed, which means we would have "commit not found" for
commits that do exist on the code host.