Rough draft of glossary
The glossary is supposed to be a sort of first-stop lookup for
"I have just encountered a term, what do I need to know about it"
for internal users/devs. This is by no means complete, it's just
what I had to mind because I'd run into it recently and needed
to know.