gitserver: More stringent check that a string is a git OID (#11597)
At a customer we are seeing revisions being returned by gitserver that are not
revisions. For example `ref: refs/heads/appsinfra/SHEP-20-review`. That string
has a length of 40 which was used by gitserver's parse HEAD logic to decide if
it found a ref. Now we use a more stringent check which should never false
positive in `quickRevParseHead` due to `: ` being present on those sorts of
lines.

There are other parts of the codebase which use the same naive check. I have put
this function in a commonly imported pkg to follow-up.