pkg/conf: suppress "type" not set in Data Center
The suppression above this commit's diff only suppresses the validation error in Server, not Data Center, when `auth.saml` or `auth.openIDConnect` lacks a `type` field. In Data Center, the `type` field is actually set to the empty string because it is serialized from config env vars.

related to https://github.com/sourcegraph/sourcegraph/issues/11148

the serialization of site config to env vars and back should be fixed as part of #11293