ensure consistent ordering of revspecs (#11656)
When reporting matching or clashing revspecs, report them in a
consistent order.  This may not matter in general, but matters
for the test, but also it seems convenient.  Added matching
test case for the multiple matching revspecs case, too.

I did not rerun this test often enough before committing it,
originally. I think I'd run that test four times, so I had about
a 94% chance of being right.

Fixes #11648.