saml: allow missing attributes (do not treat as fatal error)
If the SAML IdP returns an AuthnResponse assertion without an AttributeStatement, we still want to continue handling it. It is still useful in 2 cases:

- When the assertion has a NameID that is an email address or username (then we can use that as the email/username and don't *need* an AttributeStatement)
- When there is already a user account on Sourcegraph associated with the NameID and we can use the AuthnReponse to authenticate the client as that user

If neither of these cases holds, then the authn flow will fail later on in the flow, and we'll be able to give a better error message later on.