gitserver: Record and return last repo update and change (#11675)
We want to start recording and returning data related to when a repository
changes. This is to allow more intelligent heuristics around deciding when to
update a repository.

Git provide does not provide a useful or performant way to know when a
repository changes (outside of hooks). We could parse the output of git fetch,
but the worry is we would miss / not record a change. Instead we can hash the
state of all the references to decide this. When the hash changes we know the
repository changed. We check the hash everytime we run fetch.

Note: This change is patched cherry-picked from seebs/repoupdater. It only
includes the gitserver side changes related to recording the timestamps and
returning them. Notably it excludes the client side support for it and the
field RepoUpdateResponse.FetchInProgress.