Only use RemoteURL and EnsureRevision for ResolveRevision failing (#11694)
* all: Remove uses of EnsureRevision except for ResolveRevision

ResolveRevision is always called first, both in our browser and frontend. This
is building up to removing EnsureRevision, and instead having a more controlled
(and performant) way of triggering updates.

* gitcmd: Only fetch RemoteURL if needed

We now only set RemoteURL if resolve revision fails. Additionally, it isn't
set up front, but rather is calculated on-demand.

* changelog